Prevent map background from filling the full canvas
The cairo_renderer in recent Mapnik versions fills
the map background with the background_color
set in the active stylesheet.

By setting a clip area before calling mapnik.render()
we restrict the background filling to the area we
actually want to have the map in. Otherwise it would
fill the whole canvas and so paint over already
existing features.