support language specific config file parameters (Issue #61)
When having a locale language defined assign_if_present()
will try ot find a language specific version of the key
searched for, and only use the plain key as fallback.

E.g. with locale=de and key=name it will first check for
setting 'name.de', and will only return the value for
'name' if 'name.de' does not exist